collapsed timeClock stack into timeSheet stack, updated adminNote
